Fix condition causing missing vertical GridLines.
This patch fixes a boundary condition during vertical GridLine
calculation. The previous version used strict less-than condition. When
the min and max values fall on even multiples of the tic delta, the
absence of the GridLines is not noticeable. However, when the min and
max values do not fall on even multiples of the tic delta, the first and
last grid line are missing.

By changing the condition to use less-than-or-equal, the first and last
grid lines are included on the display.