When painting an ARC with SVGs "A" or "a" in a path, you need to set the large-arc-flag (3rd option to A or a) to 1 if you want to have an arc larger than 180°. From Mozilla's SVG documentation (https://developer.mozilla.org/en-US/docs/Web/SVG/Tutorial/Paths#Arcs):
"[...] argument is the large-arc-flag. It simply determines if the arc should be greater than or
less than 180 degrees; in the end, this flag determines which direction the arc will travel
around a given circle. "

So added simple if statements to determine if the arc would be larger than 180° and set the
parameter to 1, leaving it 0 otherwise.

Fixes  \vdobler#27.